chore!: move `get_bip34_height` to consensus.rs
This function is independent of any `ChainBackend`, it just requires the block, so I'm making it a `Consensus` associate function.

I have also changed `BlockchainInterface::update_acc` to take a block reference, as ownership isn't needed (otherwise I would need to clone the block in a PR I'm working on).
